+++ src/main/java/com/youthlink/server/member/MemberServiceImpl.java
@@ -1,0 +1,46 @@
+package com.youthlink.server.member;
+
+import com.youthlink.server.member.dto.MemberResponse;
+import com.youthlink.server.member.dto.ProfileUpdateRequest;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class MemberServiceImpl implements MemberService {
+
+    private final MemberRepository memberRepository;
+
+    @Override
+    public MemberResponse getMemberById(Long memberId) {
+        Member member = memberRepository.findById(memberId)
+                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 회원입니다. id=" + memberId));
+        return MemberResponse.from(member);
+    }
+
+    @Override
+    public MemberResponse getMemberByEmail(String email) {
+        Member member = memberRepository.findByEmail(email)
+                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 회원입니다. email=" + email));
+        return MemberResponse.from(member);
+    }
+
+    @Override
+    @Transactional
+    public MemberResponse updateProfile(Long memberId, ProfileUpdateRequest request) {
+        Member member = memberRepository.findById(memberId)
+                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 회원입니다. id=" + memberId));
+
+        member.updateProfile(
+                request.getName(),
+                request.getAge(),
+                request.getRegion(),
+                request.getEducation(),
+                request.getEmploymentStatus(),
+                request.getIncomeLevel());
+
+        return MemberResponse.from(member);
+    }
+}
